add LinksController
